Move linedelimiter to it's own pkg
Refactoring resource container
