+++ pkg/util/resourcecontainer/resource_container_linux.go
@@ -16,7 +16,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package util
+package resourcecontainer
 
 import (
 	"os"
