+++ pkg/util/resourcecontainer/resource_container_unsupported.go
@@ -1,3 +1,5 @@
+// +build !linux
+
 /*
 Copyright 2015 The Kubernetes Authors.
 
@@ -14,27 +16,16 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package util
+package resourcecontainer
 
 import (
-	"fmt"
-	"os"
+	"errors"
 )
 
-func Example_trailingNewline() {
-	ld := NewLineDelimiter(os.Stdout, "|")
-	defer ld.Flush()
-	fmt.Fprint(ld, "  Hello  \n  World  \n")
-	// Output:
-	// |  Hello  |
-	// |  World  |
-	// ||
+func RunInResourceContainer(containerName string) error {
+	return errors.New("resource-only containers unsupported in this platform")
 }
-func Example_noTrailingNewline() {
-	ld := NewLineDelimiter(os.Stdout, "|")
-	defer ld.Flush()
-	fmt.Fprint(ld, "  Hello  \n  World  ")
-	// Output:
-	// |  Hello  |
-	// |  World  |
+
+func ApplyRLimitForSelf(maxOpenFiles uint64) error {
+	return errors.New("SetRLimit unsupported in this platform")
 }
